auth/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
async-typer==0.1.8
asyncpg==0.29.0
Authlib==1.3.1
bcrypt==4.2.0
black==24.8.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-pagination==0.12.26
flake8==7.1.1
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.65.5
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
opentelemetry-api==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
protobuf==4.25.4
pika==1.3.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.8
rich==13.7.1
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
thrift==0.20.0
tomli==2.0.1
typer==0.9.4
typing_extensions==4.12.2
ua-parser-py==1.0.37
uvicorn==0.30.5
wheel==0.43.0
wrapt==1.16.0
zipp==3.20.0
itsdangerous==2.2.0
httpcore==1.0.5
httpx==0.27.0